server/requirements.txt
flask
torch
opencv-python
numpy
pillow
flask-cors
